+++ core/console/menu/helper.go
@@ -1,0 +1,5 @@
+package menu
+
+func (m *Menu) IsDaemon() bool {
+	return m.isDaemon
+}
